# feat: implement optimized PDF processing with adaptive strategies and 6x performance improvement