--- a/pyproject.toml
+++ b/pyproject.toml
@@ -50,6 +50,7 @@ dependencies = [
     "prompt-toolkit>=3.0.51",
     "pydantic>=2.11.5",
     "pypdf>=5.6.0",
+    "cryptography>=45.0.4",
 ]
 
 [project.scripts]