--- a/src/oboyu/crawler/services/__init__.py
+++ b/src/oboyu/crawler/services/__init__.py
@@ -13,6 +13,7 @@ from oboyu.crawler.services.encoding_detector import EncodingDetector
 from oboyu.crawler.services.file_discovery_service import FileDiscoveryService
 from oboyu.crawler.services.language_detector import LanguageDetector
 from oboyu.crawler.services.metadata_extractor import MetadataExtractor
+from oboyu.crawler.services.optimized_pdf_processor import OptimizedPDFProcessor
 
 __all__ = [
     "FileDiscoveryService",
@@ -20,4 +21,5 @@ __all__ = [
     "LanguageDetector",
     "EncodingDetector",
     "MetadataExtractor",
+    "OptimizedPDFProcessor",
 ]