--- a/src/oboyu/crawler/services/content_extractor.py
+++ b/src/oboyu/crawler/services/content_extractor.py
@@ -2,7 +2,10 @@
 
 import mimetypes
 from pathlib import Path
-from typing import Any, Dict, Tuple
+from typing import TYPE_CHECKING, Any, Dict, Tuple
+
+if TYPE_CHECKING:
+    import pypdf
 
 import chardet
 import charset_normalizer
@@ -260,63 +263,143 @@ class ContentExtractor:
         try:
             import pypdf
         except ImportError:
-            raise RuntimeError("pypdf library is required for PDF processing. Install with: pip install pypdf")
+            raise RuntimeError("pypdf library is required for PDF processing. Install with: uv add pypdf")
+        
+        # Check file size before processing
+        file_size = file_path.stat().st_size
+        if file_size > self.max_file_size:
+            raise RuntimeError(f"PDF file too large ({file_size / (1024*1024):.1f}MB). Maximum supported size: {self.max_file_size / (1024*1024):.1f}MB")
         
         try:
             with open(file_path, "rb") as file:
                 pdf_reader = pypdf.PdfReader(file)
                 
-                # Extract text from all pages
-                text_content = []
-                for page_num, page in enumerate(pdf_reader.pages):
-                    page_text = page.extract_text()
-                    if page_text:
-                        text_content.append(page_text)
+                # Check if PDF is encrypted and decrypt if needed
+                self._handle_pdf_encryption(pdf_reader, file_path)
                 
-                content = "\n".join(text_content)
+                # Extract text content from all pages
+                content = self._extract_pdf_text_content(pdf_reader, file_path)
                 
-                # Extract metadata
-                metadata: Dict[str, Any] = {}
-                if pdf_reader.metadata:
-                    # Extract title
-                    if pdf_reader.metadata.title:
-                        metadata["title"] = str(pdf_reader.metadata.title)
-                    
-                    # Extract creator
-                    if hasattr(pdf_reader.metadata, "creator") and pdf_reader.metadata.creator:
-                        metadata["creator"] = str(pdf_reader.metadata.creator)
-                    
-                    # Extract creation date
-                    if hasattr(pdf_reader.metadata, "creation_date") and pdf_reader.metadata.creation_date:
-                        from datetime import datetime
-                        creation_date = pdf_reader.metadata.creation_date
-                        if isinstance(creation_date, datetime):
-                            metadata["created_at"] = creation_date
-                        else:
-                            # Try to parse string date
-                            try:
-                                metadata["created_at"] = datetime.fromisoformat(str(creation_date).replace("Z", "+00:00"))
-                            except ValueError:
-                                # Store as string if parsing fails
-                                metadata["created_at"] = str(creation_date)
-                    
-                    # Extract modification date
-                    if hasattr(pdf_reader.metadata, "modification_date") and pdf_reader.metadata.modification_date:
-                        from datetime import datetime
-                        mod_date = pdf_reader.metadata.modification_date
-                        if isinstance(mod_date, datetime):
-                            metadata["updated_at"] = mod_date
-                        else:
-                            # Try to parse string date
-                            try:
-                                metadata["updated_at"] = datetime.fromisoformat(str(mod_date).replace("Z", "+00:00"))
-                            except ValueError:
-                                # Store as string if parsing fails
-                                metadata["updated_at"] = str(mod_date)
+                # Extract PDF metadata
+                metadata = self._extract_pdf_metadata(pdf_reader, len(content.split("\n\n")))
                 
                 return content, metadata
                 
         except pypdf.errors.PdfReadError as e:
-            raise RuntimeError(f"Failed to read PDF file: {e}")
+            self._handle_pdf_read_error(e, file_path)
+            raise  # This should never be reached due to exception handling
         except Exception as e:
-            raise RuntimeError(f"Failed to extract PDF content: {e}")
+            self._handle_pdf_general_error(e, file_path)
+            raise  # This should never be reached due to exception handling
+
+    def _handle_pdf_encryption(self, pdf_reader: "pypdf.PdfReader", file_path: Path) -> None:
+        """Handle PDF encryption and decryption."""
+        if pdf_reader.is_encrypted:
+            # Try to decrypt with empty password (common case)
+            if not pdf_reader.decrypt(""):
+                raise RuntimeError(f"PDF file is password-protected and cannot be processed: {file_path.name}")
+
+    def _extract_pdf_text_content(self, pdf_reader: "pypdf.PdfReader", file_path: Path) -> str:
+        """Extract text content from PDF pages with progress indication."""
+        text_content = []
+        total_pages = len(pdf_reader.pages)
+        
+        # Show progress for large PDFs (more than 10 pages)
+        if total_pages > 10:
+            print(f"Processing large PDF: {file_path.name} ({total_pages} pages)")
+        
+        for page_num, page in enumerate(pdf_reader.pages):
+            try:
+                page_text = page.extract_text()
+                if page_text and page_text.strip():
+                    text_content.append(page_text)
+                
+                # Show progress for large PDFs
+                if total_pages > 10 and (page_num + 1) % max(1, total_pages // 10) == 0:
+                    progress = ((page_num + 1) / total_pages) * 100
+                    print(f"  Progress: {progress:.0f}% ({page_num + 1}/{total_pages} pages)")
+                    
+            except Exception as page_error:
+                # Continue processing other pages if one fails
+                print(f"Warning: Failed to extract text from page {page_num + 1} of {file_path.name}: {page_error}")
+                continue
+        
+        content = "\n\n".join(text_content)  # Use double newline for better page separation
+        
+        # Warn if no text was extracted
+        if not content.strip():
+            print(f"Warning: No text content extracted from PDF: {file_path.name} (may be image-only or corrupted)")
+        
+        return content
+
+    def _extract_pdf_metadata(self, pdf_reader: "pypdf.PdfReader", extracted_pages: int) -> Dict[str, Any]:
+        """Extract metadata from PDF."""
+        metadata: Dict[str, Any] = {}
+        metadata["total_pages"] = len(pdf_reader.pages)
+        metadata["extracted_pages"] = extracted_pages
+        
+        if pdf_reader.metadata:
+            # Extract title
+            if pdf_reader.metadata.title:
+                metadata["title"] = str(pdf_reader.metadata.title)
+            
+            # Extract creator
+            if hasattr(pdf_reader.metadata, "creator") and pdf_reader.metadata.creator:
+                metadata["creator"] = str(pdf_reader.metadata.creator)
+            
+            # Extract dates
+            self._extract_pdf_dates(pdf_reader.metadata, metadata)
+        
+        return metadata
+
+    def _extract_pdf_dates(self, pdf_metadata: "pypdf.DocumentInformation", metadata: Dict[str, Any]) -> None:
+        """Extract creation and modification dates from PDF metadata."""
+        from datetime import datetime
+        
+        # Extract creation date
+        if hasattr(pdf_metadata, "creation_date") and pdf_metadata.creation_date:
+            creation_date = pdf_metadata.creation_date
+            if isinstance(creation_date, datetime):
+                metadata["created_at"] = creation_date
+            else:
+                # Try to parse string date
+                try:
+                    metadata["created_at"] = datetime.fromisoformat(str(creation_date).replace("Z", "+00:00"))
+                except ValueError:
+                    # Store as string if parsing fails
+                    metadata["created_at"] = str(creation_date)
+        
+        # Extract modification date
+        if hasattr(pdf_metadata, "modification_date") and pdf_metadata.modification_date:
+            mod_date = pdf_metadata.modification_date
+            if isinstance(mod_date, datetime):
+                metadata["updated_at"] = mod_date
+            else:
+                # Try to parse string date
+                try:
+                    metadata["updated_at"] = datetime.fromisoformat(str(mod_date).replace("Z", "+00:00"))
+                except ValueError:
+                    # Store as string if parsing fails
+                    metadata["updated_at"] = str(mod_date)
+
+    def _handle_pdf_read_error(self, error: Exception, file_path: Path) -> None:
+        """Handle PDF read errors with specific error messages."""
+        error_msg = str(error).lower()
+        if "crypt" in error_msg or "encrypt" in error_msg or "aes" in error_msg:
+            try:
+                import cryptography  # noqa: F401
+            except ImportError:
+                raise RuntimeError("PDF file requires cryptography library for decryption. Install with: uv add cryptography")
+            raise RuntimeError(f"Failed to decrypt PDF file (may be password-protected): {file_path.name}")
+        else:
+            raise RuntimeError(f"Failed to read PDF file '{file_path.name}': {error}")
+
+    def _handle_pdf_general_error(self, error: Exception, file_path: Path) -> None:
+        """Handle general PDF processing errors."""
+        error_msg = str(error).lower()
+        if "crypt" in error_msg or "aes" in error_msg:
+            try:
+                import cryptography  # noqa: F401
+            except ImportError:
+                raise RuntimeError("PDF file requires cryptography library for decryption. Install with: uv add cryptography")
+        raise RuntimeError(f"Failed to extract PDF content from '{file_path.name}': {error}")