## Summary

This PR introduces a comprehensive optimization for PDF processing in the crawler system, achieving a **6.1x average speedup** through adaptive processing strategies, parallel processing, and intelligent caching.

### Key Improvements
- **6.1x average speedup** for PDF processing across test files
- Parallel processing using ThreadPoolExecutor for multi-page documents  
- Adaptive strategy selection based on file size and page count
- Memory-efficient streaming for large files (>30MB)
- Intelligent file-hash based caching system
- Increased file size limit from 10MB to 50MB

### Performance Results
- `fastretailing_ar2024_jp.pdf`: 15.80s → 1.52s (11.9x faster)
- `marubeni_ar2024.pdf`: 43.73s → 5.12s (8.5x faster)
- Successfully processed 35/39 test files (89.7% success rate)
- Memory usage reduced by ~90% for large files through streaming

### Technical Implementation
- **Adaptive Processing Strategies**: Automatic selection between lightweight, standard, parallel, and streaming modes based on file characteristics
- **OptimizedPDFProcessor**: New processor class with strategy-based processing
- **Integration Layer**: Backward-compatible integration with existing ContentExtractor
- **Intelligent Caching**: File-hash based cache with smart key generation
- **Enhanced Error Handling**: Better error messages and progress indicators

### Files Changed
- `src/oboyu/crawler/services/optimized_pdf_processor.py` - New optimized processor implementation
- `src/oboyu/crawler/services/pdf_processor_integration.py` - Integration and compatibility layer
- `src/oboyu/crawler/services/content_extractor.py` - Updated to use optimized processor
- Updated tests and dependencies

## Test plan

- [x] All existing PDF processing tests pass
- [x] Performance testing on 39 real PDF files from test_225 directory
- [x] Memory usage testing for large files
- [x] Cache functionality testing
- [x] Error handling for corrupted/encrypted PDFs
- [x] All pre-commit hooks pass (ruff, mypy, pytest)
- [x] Backward compatibility with existing ContentExtractor interface

🤖 Generated with [Claude Code](https://claude.ai/code)